+++ tables/co_ttem_sifc.sql
@@ -2,7 +2,7 @@
 --Tabla realizada para la simulacion de una facturacion previa a la facturacion real 
 --
 CREATE TABLE CO_TTEM_SIFC(
-    TEM_SIFC_TRANS                 INT                 NOT NULL                   ,
+    TEM_SIFC_TRANS                 BIGINT              NOT NULL                   ,
     TEM_SIFC_DSKA                  INT                 NOT NULL                   ,
     TEM_SIFC_CANT                  INT                 NOT NULL                   ,
     TEM_SIFC_DCTO                  INT                 NOT NULL
