+++ tables/in_teprs.sql
@@ -4,9 +4,9 @@
 --DROP TABLE IF EXISTS in_teprs
 
 CREATE TABLE in_teprs(
-    eprs_eprs               BIGSERIAL       , 
-    eprs_dska               INT             ,
-    eprs_existencia         INT             ,
-    eprs_sede               int             ,
+    eprs_eprs               BIGSERIAL           , 
+    eprs_dska               BIGINT              ,
+    eprs_existencia         BIGINT              ,
+    eprs_sede               BIGINT              ,
 PRIMARY KEY (eprs_eprs)
 );
